+++ lib/buildPackage.nix
@@ -1,7 +1,7 @@
 { cargoBuild
 , installFromCargoBuildLogHook
 , lib
-, removeReferencesTo
+, removeReferencesToVendoredSourcesHook
 }:
 
 { cargoBuildCommand ? "cargoWithProfile build"
@@ -17,27 +17,6 @@ let
   installPhaseCommand = args.installPhaseCommand or ''
     if [ -n "$cargoBuildLog" -a -f "$cargoBuildLog" ]; then
       installFromCargoBuildLog "$out" "$cargoBuildLog"
-
-      # Strip any references to the sources directory (which may have slipped in via
-      # panic info) so depending on the binary doesn't pull in all the sources as well.
-      if [ -z "''${doNotRemoveReferencesToVendorDir-}" ] && [ -n "''${cargoVendorDir-}" ]; then
-        echo stripping references to cargoVendorDir
-
-        local allCrates=$(
-          (
-            echo "$cargoVendorDir"
-            find -L "$cargoVendorDir" -mindepth 1 -maxdepth 1 -type d | \
-              xargs -I DIR find -H DIR -type l -exec readlink '{}' \;
-          ) | sort -u
-        )
-
-        for f in $(find "$out" -type f); do
-          comm -1 -2 <(echo "$allCrates") <(strings "$f" | \
-            grep --only-matching '\(${builtins.storeDir}/[^/]\+\)' | \
-            sort -u) | \
-            xargs --verbose -I REF remove-references-to -t REF "$f"
-        done
-      fi
     else
       echo ${lib.strings.escapeShellArg ''
         !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
@@ -80,6 +59,6 @@ in
 
   nativeBuildInputs = (old.nativeBuildInputs or [ ]) ++ [
     installFromCargoBuildLogHook
-    removeReferencesTo
+    removeReferencesToVendoredSourcesHook
   ];
 })